Place nav buttons to correct positions
