+++ frontend/src/components/TaskView/FutureView/FutureViewControl.css
@@ -5,6 +5,7 @@
  */
 
 .FutureViewControl {
+  position: relative;
   display: flex;
   width: calc(100% - 6px);
   align-items: center;
@@ -25,6 +26,12 @@
   cursor: pointer;
   margin-left: 0.5em !important;
   font-size: 1.5em !important;
+}
+
+.NavButtonNDays {
+  position: absolute;
+  z-index: 1;
+  top: 250px;
 }
 
 .ControlPadding {
