+++ frontend/src/components/TaskView/FutureView/FutureViewDay.css
@@ -1,6 +1,6 @@
 .NDaysView {
   position: relative;
-  margin: -1em 0.5em;
+  margin: -1em 4px;
   background-color: #F7F7F7;
   height: 366px;
 }
